projeler/Ceren_izlenme_notifier/requirements.txt
apify-client==1.6.4
google-api-python-client==2.122.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
groq==0.37.0
tenacity==8.2.3
